// apps/frontend/src/api/models/EventTypeEntity.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * An event type object
 */
export type EventTypeEntity = {
    /**
     * The unique identifier for the event type
     */
    id: string;
    /**
     * The name of the event type
     */
    name: string;
    /**
     * A brief description of the event type
     */
    description: string | null;
};